+++ src/main/java/com/dasom/whick/entity/Pressure.java
@@ -14,6 +14,6 @@ public class Pressure {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Float pressure;
+    private Integer pressure;
     private String status;
 }
